# backend-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: EduPilot Gateway API
  version: 0.14.0
  description: |
    Hợp đồng API của gateway Go — NGUỒN SỰ THẬT (D45). Mỗi route trong mã phải có ở đây và ngược lại
    (`internal/contract` kiểm hai chiều). Route thử nằm ở `openapi.test.yaml`, chỉ có trong binary dựng bằng
    build tag `testroutes`. Phase PG chưa có endpoint nghiệp vụ; P1 (FEAT-llm-gateway US-P1-04) thêm 13 thao tác `/admin/llm/*` và `/courses/{id}/llm-budget`.
    P2 (FEAT-account-security US-P2-02) thêm 9 thao tác `/auth/{login,refresh,logout,register,verify-email,resend-verification,forgot-password,reset-password,tokens/preview}` và 4 thao tác `/me/{password,sessions,sessions/{id}}` (US-P2-04); US-P2-05 thêm 429 (`LOGIN_THROTTLED` | `RATE_LIMITED`) cho login / register / verify-email / reset-password / refresh: công khai (không Bearer); `refresh`/`logout` dùng cookie `ep_rt` (HttpOnly) + kiểm Origin.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/courses/{id}/exams/{eid}/attempts/{aid}/code/{itemId}/draft:
    put:
      operationId: saveCodeDraft
      tags: [exam]
      summary: "Lưu bản nháp mã (theo ngôn ngữ; base_rev khác rev hiện tại → 409 DRAFT_CONFLICT, không ghi đè)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: aid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Exam-Tab
          in: header
          required: true
          description: "Mã tab (uuid) sinh mỗi lần tải trang; giữ trong bộ nhớ của trang, không ở sessionStorage."
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeDraftRequest"
      responses:
        "200":
          description: Đã lưu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeDraftSaved"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/courses/{id}/exams/{eid}/attempts/{aid}/code/{itemId}/run:
    post:
      operationId: runCode
      tags: [exam]
      summary: "Chạy thử với test mẫu (Idempotency-Key bắt buộc; 10 lần / 10 phút / sinh viên; không ảnh hưởng điểm)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: aid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Exam-Tab
          in: header
          required: true
          description: "Mã tab (uuid) sinh mỗi lần tải trang; giữ trong bộ nhớ của trang, không ở sessionStorage."
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRunRequest"
      responses:
        "202":
          description: Đã xếp hàng
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunQueued"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
  /api/v1/courses/{id}/exams/{eid}/attempts/{aid}/runs/{runId}:
    get:
      operationId: getCodeRun
      tags: [exam]
      summary: "Kết quả một lần chạy thử của chính mình (chỉ test mẫu)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: aid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Kết quả
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeRunView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/courses/{id}/exams/{eid}/attempts/{aid}/code/{itemId}/submit:
    post:
      operationId: submitCode
      tags: [exam]
      summary: "Nộp lời giải (Idempotency-Key bắt buộc; nhiều lần, cách nhau 15 s, tối đa 30; lần cuối được tính điểm)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: aid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Exam-Tab
          in: header
          required: true
          description: "Mã tab (uuid) sinh mỗi lần tải trang; giữ trong bộ nhớ của trang, không ở sessionStorage."
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRunRequest"
      responses:
        "202":
          description: Đã xếp hàng
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitQueued"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/courses/{id}/exams/{eid}/attempts/{aid}/code/{itemId}/submissions:
    get:
      operationId: listCodeSubmissions
      tags: [exam]
      summary: "Lịch sử nộp của chính mình ở một câu (mới nhất trước, cursor)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: aid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Danh sách
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeSubmissionPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/courses/{id}/exams/{eid}/attempts/{aid}/submissions/{sid}:
    get:
      operationId: getCodeSubmission
      tags: [exam]
      summary: "Một bản nộp của chính mình (kèm mã nguồn; chỉ biên dịch + test mẫu, không test ẩn / trọng số / verdict chung)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: eid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: aid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Bản nộp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeSubmissionView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Chưa đăng nhập, token hết hạn / không hợp lệ / phiên bị thu hồi (`UNAUTHENTICATED`, `TOKEN_EXPIRED`, `TOKEN_INVALID`, `SESSION_REVOKED`, `INVALID_CREDENTIALS`)
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Sai vai trò, ngoài lớp hoặc Origin lạ (`FORBIDDEN`, `details.reason` = `role` | `course` | `origin`); tài khoản bị khoá (`ACCOUNT_DISABLED`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Không có route / tài nguyên / không có quyền biết (`NOT_FOUND`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Xung đột trạng thái hoặc phiên bản (`CONFLICT`, `VERSION_CONFLICT`, `IDEMPOTENCY_IN_PROGRESS`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationFailed:
      description: Dữ liệu không hợp lệ (`VALIDATION_FAILED`, `INVALID_CURSOR`, `IDEMPOTENCY_KEY_*`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Quá hạn mức (`RATE_LIMITED`, `SSE_LIMIT_REACHED`); kèm `Retry-After`
      headers:
        Retry-After:
          required: true
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unavailable:
      description: Phụ thuộc hỏng (`SERVICE_UNAVAILABLE`, `NOT_READY`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [code, message, trace_id]
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
            - BAD_REQUEST
            - UNAUTHENTICATED
            - TOKEN_EXPIRED
            - TOKEN_INVALID
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VERSION_CONFLICT
            - IDEMPOTENCY_IN_PROGRESS
            - PAYLOAD_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - VALIDATION_FAILED
            - INVALID_CURSOR
            - IDEMPOTENCY_KEY_REUSED
            - IDEMPOTENCY_KEY_REQUIRED
            - RATE_LIMITED
            - SSE_LIMIT_REACHED
            - INTERNAL
            - SERVICE_UNAVAILABLE
            - NOT_READY
            - DEADLINE_EXCEEDED
            - OVERLOADED
            - LLM_NOT_CONFIGURED
            - LLM_UNAVAILABLE
            - PROVIDER_IN_USE
            - MODEL_DIMS_MISMATCH
            - ROUTE_INVALID
            - INVALID_CREDENTIALS
            - LOGIN_THROTTLED
            - ACCOUNT_DISABLED
            - SESSION_REVOKED
            - LINK_INVALID
            - EMAIL_NOT_VERIFIED
            - JOIN_CODE_INVALID
            - COURSE_FULL
            - COURSE_ARCHIVED
            - EXAM_NOT_OPEN
            - ATTEMPT_ALREADY_SUBMITTED
            - ATTEMPT_CLOSED
            - ATTEMPT_OTHER_TAB
            - DRAFT_CONFLICT
            - EXAM_LOCKED
            - QUESTION_IN_USE
            - RESULT_NOT_PUBLISHED
            - SUBMISSION_LIMIT_REACHED
            - JUDGE_UNAVAILABLE
            - APPEAL_WINDOW_CLOSED
            - APPEAL_EXISTS
        message:
          type: string
        trace_id:
          type: string
          pattern: "^[0-9a-f]{32}$"
        details: {}
        retry_after:
          type: integer
          minimum: 1
...
    CodeDraftRequest:
      type: object
      required: [language, source, base_rev]
      additionalProperties: false
      properties:
        language:
          type: string
          enum: [c11, cpp17]
        source:
          type: string
          maxLength: 65536
        base_rev:
          type: integer
          minimum: 0
    CodeDraftSaved:
      type: object
      required: [rev, saved_at, server_time, deadline_at]
      additionalProperties: false
      properties:
        rev:
          type: integer
        saved_at:
          type: string
          format: date-time
        server_time:
          type: string
          format: date-time
        deadline_at:
          type: string
          format: date-time
    CodeRunRequest:
      type: object
      required: [language, source]
      additionalProperties: false
      properties:
        language:
          type: string
          enum: [c11, cpp17]
        source:
          type: string
          minLength: 1
          maxLength: 65536
    RunQueued:
      type: object
      required: [run_id]
      additionalProperties: false
      properties:
        run_id:
          type: string
          format: uuid
    SubmitQueued:
      type: object
      required: [submission_id]
      additionalProperties: false
      properties:
        submission_id:
          type: string
          format: uuid
    CodeSampleResult:
      type: object
      required: [name, verdict, time_ms, memory_kb]
      additionalProperties: false
      properties:
        name:
          type: string
        verdict:
          type: string
          enum: [AC, WA, TLE, MLE, RE, OLE, CE, IE]
        time_ms:
          type: integer
        memory_kb:
          type: integer
        input:
          type: string
        expected:
          type: string
        got:
          type: string
    CodeRunView:
      type: object
      required: [id, status, language, created_at, compile_ok, samples]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [QUEUED, RUNNING, DONE, ERROR, SUPERSEDED]
        language:
          type: string
        created_at:
          type: string
          format: date-time
        compile_ok:
          type: [boolean, "null"]
        compile_log:
          type: string
        samples:
          type: array
          items:
            $ref: "#/components/schemas/CodeSampleResult"
    CodeSubmissionView:
      type: object
      required: [id, status, language, created_at, compile_ok, samples, is_final]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [QUEUED, RUNNING, DONE, ERROR, SUPERSEDED]
        language:
          type: string
        created_at:
          type: string
          format: date-time
        compile_ok:
          type: [boolean, "null"]
        compile_log:
          type: string
        samples:
          type: array
          items:
            $ref: "#/components/schemas/CodeSampleResult"
        is_final:
          type: boolean
        source:
          type: string
    CodeSubmissionPage:
      type: object
      required: [items, next_cursor]
      additionalProperties: false
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/CodeSubmissionView"
        next_cursor:
          type: [string, "null"]
...